+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Parameter for EcoSystem.purgeDogu function to keep dogu config; #26
+- Option for additional purge parameters; #26
 
 ## [v1.5.1] - 2021-11-17
 ### Fixed
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Function `waitUntilAvailable(doguName, timeout)` to wait until the dogu is ready for interaction
 
 ## [v1.4.1] - 2021-09-20
-### Fixed 
+### Fixed
 - A new EcoSystem image has been released which is not compatible with the previous version.
   The disk size has increased from 64 GB to 100 GB.